[RELAND] media: Change VideoFrame's stride type to size_t
We don't support negative strides anymore, so no reason to keep it
in a signed type.

Reland of https://chromium.googlesource.com/chromium/src/+/abc3810f20082e06b47fc70fa1da35786a63dc14 with arm-chromeos
build issue fixed.

Bug: 338570700
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6027141
Commit-Queue: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1384720}

+++ media/base/mac/video_frame_mac_unittests.cc
@@ -164,8 +164,8 @@ TEST(VideoFrameMac, CorrectlyWrapsFramesWithPadding) {
     uint8_t* plane_ptr = reinterpret_cast<uint8_t*>(
         CVPixelBufferGetBaseAddressOfPlane(pb.get(), i));
     EXPECT_EQ(frame->visible_data(i), plane_ptr);
-    const int stride =
-        static_cast<int>(CVPixelBufferGetBytesPerRowOfPlane(pb.get(), i));
+    const size_t stride =
+        static_cast<size_t>(CVPixelBufferGetBytesPerRowOfPlane(pb.get(), i));
     EXPECT_EQ(frame->stride(i), stride);
     const int offset = kVisibleRectOffset / ((i == 0) ? 1 : 2);
     for (int h = 0; h < plane_size.height(); ++h) {
